refactor: Add 'isAdmin' field to UserContext interface
The 'UserContext' interface in the 'user.types.ts' file has been modified to include a new 'isAdmin' field. This field is of type boolean and is used to indicate whether the user is an admin or not. This change allows for better user role management and permissions handling in the application.