refactor: Add createAt field to Assistant and Event entities
This commit adds the `createAt` field to the `Assistant` and `Event` entities. The `Assistant` entity now has a `createAt` column with the `@CreateDateColumn()` decorator, replacing the previous `registerAt` column. Similarly, the `Event` entity's `createAt` column has been changed from a string to a `Date` type. These changes improve the consistency and clarity of the codebase.