refactor: Update initialDate and finishDate columns to use Date type
This commit updates the initialDate and finishDate columns in the Event entity to use the Date type instead of string. The columns are now set to default to the current date and time using the new Date() constructor. This change improves the consistency and accuracy of the date values stored in the database.