refactor: Update event service to use Date type for initial and finish dates
This commit updates the event service to use the Date type for the initial and finish dates when creating or updating an event. The initial date is set to the start date of the first date in the provided dates array, and the finish date is set to the end date of the last date in the array. This change improves the consistency and accuracy of the date values stored in the database.